+++ src/components/DocumentsTable.vue
@@ -3,6 +3,7 @@ import {Options, Vue} from "vue-class-component";
 import {Prop, Watch} from "vue-property-decorator";
 import {
   Action,
+  Bridgehead,
   Module,
   ProjectDocument,
   ProjectManagerContext,
@@ -17,25 +18,53 @@ import DownloadButton from "@/components/DownloadButton.vue";
 export default class DocumentsTable extends Vue {
   @Prop() readonly context!: ProjectManagerContext;
   @Prop() readonly projectManagerBackendService!: ProjetManagerBackendService;
-  @Prop() readonly module!: Module;
-  @Prop() readonly action!: Action;
-  @Prop() readonly projectDocuments!: ProjectDocument[];
+  @Prop() readonly downloadAction!: Action;
+  @Prop() readonly fetchListAction!: Action;
   @Prop() readonly iconClass: string | undefined = undefined;
   @Prop() readonly text!: string;
+  @Prop() readonly bridgeheads!: Bridgehead[];
+
   canDownload = false;
+  Module = Module;
+  projectDocuments: ProjectDocument[] = [];
+  projectDocumentIds = new Set<string>();
 
   @Watch('projectManagerBackendService', {immediate: true, deep: true})
   onContextChange(newValue: ProjetManagerBackendService, oldValue: ProjetManagerBackendService) {
-    this.updateCanDownload()
+    this.updateCanDownload();
   }
 
   async created() {
     this.updateCanDownload()
   }
 
-  updateCanDownload() {
-    this.projectManagerBackendService.isModuleActionActive(this.module, this.action).then(result => this.canDownload = result)
+  fetchProjectDocuments() {
+    this.projectDocuments = [];
+    this.projectDocumentIds = new Set<string>();
+    this.bridgeheads.forEach(bridgehead => this.projectManagerBackendService
+        .fetchData(Module.PROJECT_DOCUMENTS_MODULE, this.fetchListAction, this.createContext(bridgehead), new Map())
+        .then(results => (results as ProjectDocument[]).forEach(result => {
+          let key = JSON.stringify(result);
+          if (!this.projectDocumentIds.has(key)){
+            this.projectDocuments.push(result);
+            this.projectDocumentIds.add(key);
+          }
+        })));
   }
+
+  updateCanDownload() {
+    this.projectManagerBackendService.isModuleActionActive(Module.PROJECT_DOCUMENTS_MODULE, this.downloadAction).then(result => {
+      this.canDownload = result;
+      if (this.canDownload) {
+        this.fetchProjectDocuments();
+      }
+    })
+  }
+
+  createContext(bridgehead: Bridgehead) {
+    return new ProjectManagerContext(this.context.projectCode, bridgehead.bridgehead);
+  }
+
 
 }
 </script>
@@ -47,30 +76,30 @@ export default class DocumentsTable extends Vue {
     <table>
       <thead>
       <tr>
+        <th>Label</th>
         <th>Original Filename</th>
         <th>URL</th>
         <th>Created At</th>
         <th>Bridgehead</th>
         <th>Creator Email</th>
-        <th>Label</th>
         <th>Type</th>
         <th>Download</th>
       </tr>
       </thead>
       <tbody>
       <tr v-for="(projectDocument, index) in projectDocuments" :key="index">
+        <td>{{ projectDocument.label }}</td>
         <td>{{ projectDocument.originalFilename }}</td>
-        <td>{{ projectDocument.url }}</td>
+        <td><a :href="projectDocument.url">{{ projectDocument.url }}</a></td>
         <td>{{ projectDocument.createdAt }}</td>
         <td v-if="projectDocument.bridgehead != 'NONE'">{{ projectDocument.bridgehead }}</td>
         <td v-if="projectDocument.bridgehead === 'NONE'"></td>
         <td>{{ projectDocument.creatorEmail }}</td>
-        <td>{{ projectDocument.label }}</td>
         <td>{{ projectDocument.type }}</td>
         <td>
           <DownloadButton v-if="canDownload && projectDocument.originalFilename"
                           :context="context" :project-manager-backend-service="projectManagerBackendService"
-                          :module="module" :action="action" :icon-class="iconClass"
+                          :module="Module.PROJECT_DOCUMENTS_MODULE" :action="downloadAction" :icon-class="iconClass"
                           :filename="projectDocument.originalFilename"/>
         </td>
       </tr>